Move type serialization to "frame" module
The types are part of the CQL binary protocol so let's move types.rs to
the "frame" module.
