+++ scylla/src/frame/mod.rs
@@ -1,5 +1,6 @@
 pub mod request;
 pub mod response;
+pub mod types;
 
 use anyhow::Result;
 use bytes::{Buf, BufMut, Bytes};
