+++ scylla/src/lib.rs
@@ -6,4 +6,3 @@ extern crate anyhow;
 pub mod frame;
 pub mod query;
 pub mod transport;
-pub mod types;
